+++ application/views/History_view.php
@@ -1,0 +1,17 @@
+<h2>History Transactions</h2>
+{pagination}
+<table class="table">
+<thead>
+  <tr>
+    <th><a href="history?sort=id">Transaction ID</a></th>
+    <th><a href="history?sort=type">Transaction Type</a></th>
+    <th><a href="history?sort=part_id">Part ID</a></th>
+    <th><a href="history?sort=robot_id">Robot ID</a></th>
+    <th><a href="history?sort=amount">Amount</a></th>
+    <th><a href="history?sort=time">Transaction Time</a></th>
+  </tr>
+</thead>
+<tbody>
+    {display_history}
+</tbody>
+</table>