+++ changelog.md
@@ -6,6 +6,41 @@
 - Jason Chan     (jasonlearn)
 
 # Changelog
+## [0.6.0] - 2017-04-2
+### New Features
+- Added Sortable History page
+### Changed Features
+- Fixed incorrect transaction types on robot sale and part recycling
+- Tied up loose ends
+
+## [0.5.7] - 2017-04-2
+### Changed Features
+- fixed assembly page images
+
+## [0.5.6] - 2017-04-2
+### New Features
+- Added role selection
+- Menubar items disapear based on role
+
+## [0.5.5] - 2017-04-2
+### Changed Features
+- Shipping robots creates transactions
+- added commenting to functions
+
+## [0.5.4] - 2017-04-2
+### New Features
+- Added assembly page functionality with database
+- Assembly page can now build robots
+- Assembly page can now return parts
+- Assembly page, robots tab can now sell robots
+### Changed Features
+- Fixed bug with Umbrella_API
+
+## [0.5.3] - 2017-04-2
+### New Features
+- Added updating of the transactions for parts purchases
+- Parts page now sorted by line
+
 ## [0.5.2] - 2017-04-02
 ### New Features
 - Added Sell tab on Manage page